requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.2.1
click==8.3.3
fastapi==0.136.0
h11==0.16.0
idna==3.13
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
passlib==1.7.4
pydantic==2.13.3
pydantic-core==2.46.3
python-dotenv==1.2.2
python-multipart==0.0.17
starlette==1.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.45.0
